+++ driver/src/main/kotlin/io/github/octaviusframework/driver/converter/EnumConverters.kt
@@ -6,10 +6,10 @@ import io.github.octaviusframework.driver.converter.parameter.mapper.ParameterCo
 import io.github.octaviusframework.driver.converter.parameter.mapper.SerializationContext
 import io.github.octaviusframework.driver.converter.result.mapper.DeserializationContext
 import io.github.octaviusframework.driver.converter.result.mapper.ResultConverter
-import io.github.octaviusframework.identifier.CaseConvention
-import io.github.octaviusframework.driver.identifier.CaseConverter
 import io.github.octaviusframework.driver.identifier.QualifiedName
 import io.github.octaviusframework.driver.type.PgType
+import io.github.octaviusframework.identifier.CaseConvention
+import io.github.octaviusframework.identifier.CaseConverter
 import kotlin.reflect.KClass
 import kotlin.reflect.KType
 
@@ -87,4 +87,3 @@ internal class EnumResultConverter<T : Enum<T>>(
             ?: throw IllegalArgumentException("Unknown enum value: $source for enum ${enumClass.simpleName}")
     }
 }
-